+++ backend/gce.go
@@ -643,7 +643,16 @@ func (p *gceProvider) apiRateLimit(ctx gocontext.Context) error {
 		}
 
 		// Sleep for up to 1 second
+		var span *trace.Span
+		if trace.FromContext(ctx) != nil {
+			ctx, span = trace.StartSpan(ctx, "GCE.timeSleep.apiRateLimit")
+		}
+
 		time.Sleep(time.Millisecond * time.Duration(mathrand.Intn(1000)))
+
+		if trace.FromContext(ctx) != nil {
+			span.End()
+		}
 	}
 }
 
@@ -1074,7 +1083,9 @@ func (p *gceProvider) stepWaitForInstanceIP(c *gceStartContext) multistep.StepAc
 		State:   ProgressNeutral,
 	})
 
+	ctx, span = trace.StartSpan(ctx, "GCE.timeSleep.WaitForInstanceIP")
 	time.Sleep(p.bootPrePollSleep)
+	span.End()
 
 	zoneOpCall := p.client.ZoneOperations.Get(p.projectID, c.zoneName, c.instanceInsertOpName).Context(c.ctx)
 
@@ -1149,7 +1160,11 @@ func (p *gceProvider) stepWaitForInstanceIP(c *gceStartContext) multistep.StepAc
 		}).Debug("sleeping before checking instance insert operation")
 
 		c.progresser.Progress(&ProgressEntry{Message: ".", Raw: true})
+
+		var span *trace.Span
+		ctx, span = trace.StartSpan(ctx, "GCE.timeSleep.afterInstanceInsertCompletion")
 		time.Sleep(p.bootPollSleep)
+		span.End()
 	}
 }
 
@@ -1612,7 +1627,11 @@ func (i *gceInstance) UploadScript(ctx gocontext.Context, script []byte) error {
 			}
 
 			i.progresser.Progress(&ProgressEntry{Message: ".", Raw: true})
+			var span *trace.Span
+			ctx, span = trace.StartSpan(ctx, "GCE.timeSleep.uploadRetry")
 			time.Sleep(i.provider.uploadRetrySleep)
+			span.End()
+
 		}
 	}()
 
@@ -1822,7 +1841,11 @@ func (i *gceInstance) stepWaitForInstanceDeleted(c *gceInstanceStopContext) mult
 		"duration": i.ic.StopPrePollSleep,
 	}).Debug("sleeping before first checking instance delete operation")
 
+	var span *trace.Span
+	ctx := c.ctx
+	ctx, span = trace.StartSpan(ctx, "GCE.timeSleep.WaitForInstanceDeleted")
 	time.Sleep(i.ic.StopPrePollSleep)
+	span.End()
 
 	zoneOpCall := i.client.ZoneOperations.Get(i.projectID,
 		i.zoneName, c.instanceDeleteOp.Name)